+++ Assets/Scripts/Bars.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 026f27d5f6630b34e8a7a679ea84eba3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
