+++ Assets/Scripts/Bars/Bars.asmdef
@@ -1,8 +1,10 @@
 {
-    "name": "Utilities",
+    "name": "Bars",
     "rootNamespace": "",
     "references": [
-        "GUID:560b04d1a97f54a4e82edc0cbbb69285"
+        "GUID:560b04d1a97f54a4e82edc0cbbb69285",
+        "GUID:59fd921a24d1a52408da002aba99c103",
+        "GUID:a9a89846cc8daa8478fb2a56419e7f4c"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
