+++ Assets/Scripts/Bars/Bars.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2e8679054c717c143852b880d782cb14
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
