+++ Assets/Scripts/Data/Data.asmdef
@@ -1,8 +1,9 @@
 {
-    "name": "Utilities",
+    "name": "Data",
     "rootNamespace": "",
     "references": [
-        "GUID:560b04d1a97f54a4e82edc0cbbb69285"
+        "GUID:560b04d1a97f54a4e82edc0cbbb69285",
+        "GUID:59fd921a24d1a52408da002aba99c103"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
