+++ Assets/Scripts/Data/Data.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a9a89846cc8daa8478fb2a56419e7f4c
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
