+++ Assets/Scripts/Utilities/LoadSpriteData.cs.meta
