+++ Assets/Scripts/Utilities/SingletonMonoBehaviour.cs.meta
