+++ Assets/Scripts/Works/MiniGameManager.cs.meta
