+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,18 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/MiniGame.unity
+    path: Assets/Scenes/TitleScene.unity
+    guid: 83d5a4cbbb05c704b8221d531ed69fa6
+  - enabled: 1
+    path: Assets/Scenes/MiniGameScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
+  - enabled: 1
+    path: Assets/Scenes/BarScene.unity
+    guid: 33599d8288972d94bbffd5af0f21e0d5
+  - enabled: 1
+    path: Assets/Scenes/HomeScene.unity
+    guid: 40929f7e0c2d7884899e4f4d76f189d7
+  - enabled: 1
+    path: Assets/Scenes/ResultScene.unity
+    guid: 4c980412cdd2cc345a11801a26db5590
   m_configObjects: {}
